+++ packages/suite-native/src/reducers/store.ts
@@ -9,17 +9,20 @@ import { createLogger } from 'redux-logger';
 import suiteMiddlewares from '@suite-middlewares/index';
 import walletMiddlewares from '@wallet-middlewares/index';
 import settingsMiddlewares from '@settings-middlewares/index';
+import firmwareMiddlewares from '@firmware-middlewares/index';
 
 import suiteReducers from '@suite-reducers/index';
 import onboardingReducers from '@onboarding-reducers/index';
 import walletReducers from '@wallet-reducers/index';
 import settingsReducers from '@settings-reducers/index';
+import firmwareReducers from '@firmware-reducers/index';
 
 const reducers = combineReducers({
     ...suiteReducers,
     onboarding: onboardingReducers,
     wallet: walletReducers,
     settings: settingsReducers,
+    firmware: firmwareReducers,
 });
 
 export type AppState = ReturnType<typeof reducers>;
@@ -29,6 +32,7 @@ const middlewares = [
     ...suiteMiddlewares,
     ...walletMiddlewares,
     ...settingsMiddlewares,
+    ...firmwareMiddlewares,
 ];
 
 const enhancers: any[] = [];
