+++ packages/suite/src/middlewares/firmware/index.ts
@@ -1,0 +1,3 @@
+import firmware from './firmwareMiddleware';
+
+export default [firmware];
