+++ packages/suite/src/reducers/firmware/firmwareReducer.ts
