+++ packages/suite/src/reducers/firmware/index.ts
@@ -1,0 +1,3 @@
+import firmware from './firmwareReducer';
+
+export default firmware;
