+++ packages/suite/src/reducers/store.ts
@@ -7,17 +7,20 @@ import suiteMiddlewares from '@suite-middlewares';
 import walletMiddlewares from '@wallet-middlewares';
 import onboardingMiddlewares from '@onboarding-middlewares';
 import settingsMiddlewares from '@settings-middlewares';
+import firmwareMiddlewares from '@firmware-middlewares';
 
 import suiteReducers from '@suite-reducers';
 import walletReducers from '@wallet-reducers';
 import onboardingReducers from '@onboarding-reducers';
 import settingsReducers from '@settings-reducers';
+import firmwareReducers from '@firmware-reducers';
 
 const rootReducer = combineReducers({
     ...suiteReducers,
     onboarding: onboardingReducers,
     wallet: walletReducers,
     settings: settingsReducers,
+    firmware: firmwareReducers,
 });
 
 export type AppState = ReturnType<typeof rootReducer>;
@@ -28,6 +31,7 @@ const middlewares = [
     ...walletMiddlewares,
     ...onboardingMiddlewares,
     ...settingsMiddlewares,
+    ...firmwareMiddlewares,
 ];
 
 const enhancers: any[] = [];
