+++ packages/suite/src/reducers/suite/index.ts
@@ -3,7 +3,6 @@ import suite from './suiteReducer';
 import devices from './deviceReducer';
 import modal from './modalReducer';
 import log from './logReducer';
-import firmware from './firmwareReducer';
 import notifications from './notificationReducer';
 import resize from './resizeReducer';
 
@@ -13,7 +12,6 @@ export default {
     modal,
     devices,
     log,
-    firmware,
     notifications,
     resize,
 };
