+++ packages/suite/src/types/firmware/index.ts
@@ -1,4 +1,3 @@
 import { FirmwareActions as FirmwareActions$ } from '@firmware-actions/firmwareActions';
 
 export type FirmwareActions = FirmwareActions$;
-
