+++ packages/suite/src/views/onboarding/steps/Firmware/Container.ts
@@ -2,7 +2,7 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { injectIntl, WrappedComponentProps } from 'react-intl';
 
-import * as firmwareUpdateActions from '@suite/actions/firmware/firmwareActions';
+import * as firmwareUpdateActions from '@firmware-actions/firmwareActions';
 import * as onboardingActions from '@onboarding-actions/onboardingActions';
 import { Dispatch, AppState } from '@suite-types';
 import Step from './index';
@@ -11,7 +11,6 @@ const mapStateToProps = (state: AppState) => ({
     deviceCall: state.onboarding.deviceCall,
     device: state.suite.device,
     firmwareUpdate: state.firmware,
-    // locks: state.suite.locks,
 });
 
 const mapDispatchToProps = (dispatch: Dispatch) => ({
